requirements.txt
pyopengl
pygame
PyOpenGL-accelerate